+++ docs/user_guide/changelog.rst
@@ -35,6 +35,7 @@ API changes
 - Default mean reduction strategies have changed from the deprecated PostAndLoss strategy to Post or Running
   based on optimiser accumulation type
 - Mean reduction strategy can now be set via ``poptorch.Options.Training.setMeanAccumulationAndReplicationReductionStrategy``.
+- Add warning that IPU-specific optimiser states cannot be read from the host, when calling ``get_state()`` on poptorch.optim optimisers
 
 v2.2 (Poplar SDK 2.2)
 =====================